+++ spec/data/models.rb
@@ -10,6 +10,15 @@ class Post < BasePost
   translates :image_file_name, :image_content_type, :image_file_size, :image_updated_at, :image_fingerprint
 end
 
+class PostWithStyles < BasePost
+  has_attached_file :image,
+    url:    '/system/:test_env_number/:class/:attachment/:id/:locale/:style-:fingerprint.:extension',
+    styles: {thumb: '10x10', large: '40x40'}
+  validates_attachment :image, content_type: {content_type: ['image/png']}
+
+  translates :image_file_name, :image_content_type, :image_file_size, :image_updated_at, :image_fingerprint
+end
+
 class OnlyProcessPost < BasePost
   has_attached_file :image,
     url:          '/system/:test_env_number/:class/:attachment/:id/:locale/:style-:fingerprint.:extension',