+++ CHANGELOG.md
@@ -4,7 +4,7 @@
     * Drop support for ruby < 2.4.6
     * Drop support for globalize < 5.3   
 * Bugfixes
-    * (none)
+    * Fix `attachment.clear(:style)` not deleting the files ([#17](https://github.com/emjot/paperclip-globalize3/pull/17)) 
 * Enhancements
     * Allow globalize 5.3
       With globalize 5.3.0 paperclip-globalize3 is again compatible witj Rails 4.2